requirements.txt
flask
rembg
pillow
onnxruntime
flask-cors